Re-reverse margins for questions
This reverts the changes originally made here for IE8 compatibility:
https://github.com/alphagov/digitalmarketplace-frontend-toolkit/commit/cc9c80010c13440a1dd36cdce4b0f4517f328fac

The "no visual changes" comment on the pull-request for that commit was wishful thinking, as it turns out this did  have an impact on how questions are displayed, making the question text out of alignment with the question number.

The current version would work *if* the opening paragraph was the first child of
its containing span.  The problem is that our questions render with structure as follows
- the first paragraph is the *second* element in its container (after the question number span), so doesn't pick up the "first-child" pseudo selector.

e.g.
```
<span class="question-heading">
  <span class="question-number">1</span>
  <p>Do you sell onions?</p>
</span>
```

This new version is again not compatible with IE8, as "last-child" will not work there.