Users should by able to see the full list of filters
User story
----------------------
As a user, when I'm searching through a list
I need to be able to see all the filters that I have selected
So that I can tell how I have refined my search results

**The people required for kick off are:**

- Laura + developer + Jeremy

**Context/Problem**

The height of the filter boxes is fixed and this means if the list of filters is above 6, the filters are hidden and users have to scroll down to see the full list. On the DOS opportunities page there is no search summary so if users select filters they won't be able to refer back to which ones they have selected easily.

**What are we making?**

- changing the height of the filter boxes so it is no longer restricted and  expands to include the full range of filters
- filters already behave this way on mobile

**Info/Consideration**
- @jeremy will update the screenshots of the user flow

Ticket: https://trello.com/c/9oSx43Mv/219-users-should-by-able-to-see-the-full-list-of-filters

Bump version to 25.1.0