Move shim links functionality into the toolkit
This functionality has been taken from the buyer FE and moved
into the FE toolkit as it is currently duplicated across multiple
apps and can instead be stored in a central location.

There are three minor changes to this from what already existed
in the apps.

- Improved inline documentation
- Semi colon at the beginning of the file. GOV.UK FE toolkit does
not add trailing semi colons and as this will now be imported
higher up in application.js files with external dependancies the
trailing semi colon will not have been taken care of yet.
- Remove unused customisation of shim links functionality. Before
v 5.0.1 of the GOVUK FE toolkit, it was possible to add or
override internals of `GOVUK.shimLinksWithButtonRole` by passing
them into `init`. We were using this for "link-buttons". However,
now we no longer have "link-buttons" and we have upgraded past
v5.0.1 of the GOV.UK FE toolkit so this doesn't functionality
doesn't even exist anymore. See GOV.UK FE toolkit commit for
further details:
https://github.com/alphagov/govuk_frontend_toolkit/commit/e882086f8e1ed00ee86c7b276b477b498ca6035a