Fix bug which would show empty question-number div
If `kwargs.question_number` has not been set then it is not
defined. However if a variable is not defined then it will
evaluate `True` in the statement `if kwargs.question_number is not none`.
Therefore, we if don't define `kwargs.question_number` then we
accidently show an empty `div.question-number` when we shouldn't.