Move supplier question macros to toolkit
There are some macros in the supplier frontend which wrap the templates.

We already have some macros which do a similar thing but are designed to
use question objects generated by the content loader.

These are useful for when we don't have question objects, but still want
a convenient way to generate inputs.

Moving them to the toolkit allow them to be used by other apps if
desired.