Extend `_selection-button` with conditionally revealing content
We want to use the pattern defined in GovUK pattern library for
conditionally revealing content
(http://govuk-elements.herokuapp.com/form-elements/#form-toggle-content).

This allows the options for radio selection buttons to have an extra
attribute `reveal` which defines the text box to reveal.

Laurence has approved the validation styling for the revealed component.