Bump version to 28.8.2
Given we had dropped IE6 support (https://github.com/alphagov/digitalmarketplace-frontend-toolkit/pull/396)
, we no longer need to support IE6. We had some leftover
references to `mobile-ie6` which I don't think we need as
this variable would only come into play when also supporting
IE6 and is not needed for other versions of IE. This is based
on looking at how the GOV.UK elements does not define `mobile-ie6`
for IE7/8 stylesheets and be looking at the conditionals code
that would use that variable. I believe any variations such as
`mobile-ie7` or `mobile-ie8` are not variables to use and would do
nothing.

https://github.com/alphagov/govuk_elements/tree/master/assets/sass
https://github.com/alphagov/govuk_frontend_toolkit/blob/master/stylesheets/_conditionals.scss