Bump version to 28.9.1
Removes code documenation and published documentation support
for IE7. Clarifies that it is only IE8 that requires targetted
CSS support (and therefore we don't need a specific IE9
stylesheet in our published documentation).

Note, this doesn't actually drop support from IE7 when pulled
into an application. It's only if an application drops its
IE7 specific stylesheet that it would lose support for IE7, either
before or after pulling this commit in. This commit just means
that it no longer looks like we support IE7 generally on the
Marketplace.