Fix `jasmineVerison` in `gulp-jasmine-phantom` options
`gulp-phantom-jasmine` takes an object of options with initialized. One
of these in the version of Jasmine to use (it supports v2.0 - v2.4). To
set this, the key used should be `jasmineVersion`, not `jasmine`. If
`gulp-phantom-jasmine` doesn't find this key, it defaults to 2.0, so
this update doesn't actually make a functional difference.