Add `banner` option to temporary messages
Temporary messages had no margin at the top. This makes sense if in a
one-third column, or if there's nothing above it.

When used as a full width banner message however, if there is something
above it - such as a flash message - then there is no spacing between
them and it looks weird.

Flash messages use 15px of margin-top, so I've used the same value here.