Add an optional "finally, ..." clause to selection buttons
In a number of our forms (c.f. in buyer frontend choose-lot.html,
save-search.html)
we have the pattern where the final radio button is prefaced by a line
reading "or", to signpost to the reader that it is an option unlike the
others.

Currently the forms that do this do so by hand coding the form, rather
than using the form template, or by mushing together several form
templates.

By adding this feature to the form template, we can greatly simplify the
places where this pattern is used.