Add an id to hidden fields so that the labels match correctly
For fields which are hidden behind a selection button, the label was
associated with the div tag rather than the input tag.

This causes an issue with functional tests and also could cause an issue
with accessability, so this commit changes the template so that the
input field has an additional unique id that is used by it and the
label.