Fix usablity issue with removing entries
Currently the the "Remove" button is shown to all list entries except for
the first entry. This causes a problem if the user would like to remove
the first option. I understand that we might not want the user to be able
to remove all list entries but instead the logic should have been

Display the remove button against each list entry unless it is the only
entry in the list then remove it.

This issue was raised by a failing test after I removed Hogan.js
I wanted to keep this as a seperate commit so that other developers can
see why this change was made.