Move "Built by" link in footer to separate line
We want our footer to look more like the footer from GOV.UK and GOV.UK
Deisgn System, so we can swap the components on an app by app basis in
the future.

This commit moves the text "Built by GDS" to a separate line, similar to
how it is styled in alphagov/digitalmarketplace-govuk-frontend.