Change single field pricing question to also look for maximum price
Before DOS5 all pricing questions with a single field were on G-Cloud
and the field was `price`. In DOS5 we switched to asking for only
a specialist's maximum day rate, rather than both a maximum and
minimum rate. This means we now have single field pricing questions
with fields `price` and `maximum_price`.

The field name is used to populate the form input name, for
questions with `maximum_price` this attribute was empty. This
caused a bug in the admin-frontend where an admin user could not
update a DOS5 specialists maximum day rate.

Update the name attribute on single field questions to look for
`maximum_price` if `price` is not present. I initially felt a bit
bad about hard-coding the second option, but logically it does
make sense that a single field pricing question is either asking
for a service's price (G-Cloud) or maximum price (DOS specialists).
I have checked the other questions in DOS and G-Cloud and verified
that these are the only two fields that appear on their own.

https://trello.com/c/XGAEqKRu/2165-ccs-category-admins-cant-add-new-specialist-types-to-digital-specialists-services-on-dos5